Actualizacion de Clase Transporte y diagrama de clases
Agregamos el atributo idTransporte, y modificamos la clase de Terminal agregando el metodo de calcular precio; a su vez eliminando calcularTarifa dentro de Viaje

+++ SUBE/src/datos/Subte.java
@@ -20,7 +20,7 @@ public class Subte extends Transporte {
 		this.letra = letra;
 	}
 
-	public float calcularMonto() {
+	public double calcularMonto() {
 			DatosGenerales datosGenerales = new DatosGenerales();
 			return datosGenerales.getMontoSubte();
 	}
